requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
python-telegram-bot==20.7